requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
streamlit>=1.31.0
httpx>=0.26.0
pyperclip>=1.8.2
